mail_service/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
configparser==4.0.2
contextlib2==0.6.0.post1
Django==1.11.29
django-celery==3.3.1
django-celery-fulldbresult==0.5.3
django-forms-bootstrap==3.1.0
djangorestframework==3.9.4
importlib-metadata==2.1.3
kombu==3.0.37
pathlib==1.0.1
pathlib2==2.3.7.post1
psycopg2==2.8.6
python-dotenv==0.18.0
pytz==2023.3
redis==2.10.6
scandir==1.10.0
six==1.16.0
typing==3.10.0.0
vine==1.3.0
zipp==1.2.0
